pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling==1.27.0"]

[dependency-groups]
dev = [
  "black[jupyter]==25.1.0",
  "cibuildwheel==2.23.0",
  "coverage==7.6.12",
  "creosote==4.0.1",
  "debugpy==1.8.13",
  "isort==6.0.1",
  "lefthook==1.11.3",
  "linkify-it-py==2.0.3",
  "loguru-mypy==0.0.4",
  "mdformat-black==0.1.1",
  "mdformat-footnote==0.1.1",
  "mdformat-frontmatter==2.0.8",
  "mdformat-gfm==0.4.1",
  "mdformat-pyproject==0.0.2",
  "mdformat-shfmt==0.2.0",
  "mdformat-tables==1.0.0",
  "mdformat==0.7.22",
  "mypy==1.15.0",
  "nbconvert==7.16.6",
  "packaging==24.2",
  "pandas-stubs==2.2.3.250308",
  "pydantic==2.10.6",
  "pydeps==3.0.1",
  "pylint==3.3.5",
  "pytest-profiling==1.8.1",
  "pytest-xdist==3.6.1",
  "pytest==8.3.5",
  "ruff==0.9.10",
  "ssort==0.14.0",
  "tinycss2==1.4.0",
  "toml-sort==0.24.2",
  "types-openpyxl==3.1.5.20241225",
  "types-requests==2.32.0.20241016",
  "types-tqdm==4.67.0.20250401",
]

[project]
authors = [
  {email = "Sander.Maijers@han.nl", name = "Sander Maijers"},
]
classifiers = [
  "Environment :: Console",
  "Framework :: FastAPI",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "docling-core[chunking]==2.44.2",
  "docling==2.26.0",
  "httpx[http2]==0.28.1",
  "openai==1.66.3",
  "pydantic==2.10.6",
  "python-multipart==0.0.20",
  "qdrant-client==1.13.3",
  "sentence-transformers==3.4.1",
  "spacy==3.8.7",
  "tqdm==4.67.1",
  "uvicorn==0.35.0",
]
description = "Knowledge Platform Management App library"
dynamic = ["version"]
license = {text = "NONE"}
name = "knowledgeplatformmanagement-generic"
readme = "README.md"
requires-python = ">=3.12"

[tool.pytest.ini_options]
pythonpath = ["src", "src/knowledgeplatformmanagement_han", "src/knowledgeplatformmanagement_generic"]
addopts = ["--import-mode=importlib"]
testpaths = ["tests"]
# keep your existing pythonpath:
# pythonpath = ["src", "src/knowledgeplatformmanagement_han", "src/knowledgeplatformmanagement_generic"]

